# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Manga-Store"
host: "localhost:8080"
tags:
  - name: "user"
    description: "Operations about user"
paths:
  /usuarios:
    post:
      tags:
        - "user"
      summary: "Create user"
      operationId: "createUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created Usuario object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
    get:
      tags:
        - "user"
      summary: "Find all users"
      operationId: "findAllUsers"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
    put:
      tags:
        - "user"
      summary: "Update user"
      operationId: "UpdateUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Update Usuario object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
  /usuarios/manga/{cpf}:
    post:
      tags:
        - "user"
      summary: " Add Manga on User"
      operationId: "AddMangaUser"
      produces:
        - "application/json"
      parameters:
        - name: "cpf"
          in: "path"
          description: "CPF of Usuario"
          required: true
          type: "integer"
          minimum: 1.0
        - name: "body"
          in: "body"
          description: "Add manga on Usuario object"
          required: true
          schema:
            $ref: "#/definitions/Manga"
      responses:
        default:
          description: "successful operation"

definitions:
  User:
    type: "object"
    properties:
      name:
        type: "string"
      cpf:
        type: "string"
      login:
        type: "string"
      senha:
        type: "string"
      email:
        type: "string"
      telefone:
        type: "string"
      mangas:
        type: "array"
        items:
          $ref: "#/definitions/Manga"
  Manga:
    type: "object"
    properties:
      name:
        type: "string"
      version:
        type: "string"
      description:
        type: "string"
      publishCompany:
        type: "string"
      price:
        type: "number"
      status:
        type: "string"